än yy tee nyt. Valmis

+++ controllers/blogs.js
@@ -1,36 +1,42 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/users')
+const { MODE } = require('../utils/config')
 const { info } = require('../utils/logger')
+const { userExtractor } = require('../utils/middleware')
+
+
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({})
+        .populate('user', { username: 1, name: 1, id: 1 })
     response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-    const body = await request.body
+blogsRouter.post('/', userExtractor, async (request, response) => {
+    const body = request.body
     if (!body.title && !body.url) {
         response.status(400).end()
         return
     }
-    info('userId 123: ', body.userId)
-    const user = await User.findById(body.userId)
+    const user = await User.findById(request.user)
     const blog = new Blog({ ...body, user: user.id })
-    info('user123: ', user)
+
 
     const savedBlog = await blog.save()
-    info('savedBlog12: ', savedBlog.toJSON())
-    const blogJson = savedBlog.toJSON()
-    user.blogs.concat(blogJson.id)
+    const str = JSON.stringify(savedBlog)
+    user.blogs = user.blogs.concat(JSON.parse(str).id)
     await user.save()
 
     response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+    const blog = await Blog.findById(request.params.id)
+    if (request.user === blog.user) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    }
 })
 
 blogsRouter.put('/:id', async (requst, response) => {
